swiftfood/requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
croniter==0.3.30
DateTime==4.3
Django==2.2.4
django-celery-beat==1.5.0
django-celery-monitor==1.1.2
django-celery-results==1.0.4
django-debug-toolbar==2.0
django-extensions==2.2.1
django-filter==2.2.0
django-guardian==2.0.0
django-model-utils==3.2.0
django-rest-swagger==2.2.0
django-rq==1.2.0
django-rq-scheduler==1.1.1
django-timezone-field==3.0
django-uuidfield==0.5.0
djangorestframework==3.9.2
idna==2.8
importlib-metadata==0.23
itypes==1.1.0
Jinja2==2.10.1
kombu==4.6.5
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
openapi-codec==1.3.2
Pillow==6.1.0
python-crontab==2.3.9
python-dateutil==2.8.0
pytz==2019.2
redis==3.3.8
requests==2.22.0
rq==1.1.0
rq-scheduler==0.9.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
zipp==0.6.0
zope.interface==4.7.1
